pyproject.toml
[project]
name = "wordle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.5",
    "pygame-ce>=2.5.3",
    "rich>=14.0.0",
    "typing>=3.10.0.0",
]
